requirements.txt
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
youtube-data-api==0.6.0
requests==2.31.0
pandas==2.1.4
numpy==1.24.3
plotly==5.17.0
dash==2.14.2
dash-bootstrap-components==1.5.0
textblob==0.17.1
vaderSentiment==3.3.2
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
python-dotenv==1.0.0
schedule==1.2.0
tqdm==4.66.1
wordcloud==1.9.2
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2
transformers==4.35.2
torch==2.1.1
